PlayerSpawn CustomInspector
Began adding a custom inpsector to the PlayerSpawn script - this now hides the scene selection if this scene is the first in the level.

Need to see if there's away to not override the variables I don't want to hide.

Hide Values in Inspector: https://answers.unity.com/questions/192895/hideshow-properties-dynamically-in-inspector.html